Fix circular dependency, and add typedef docs
