+++ src/utils/tasks.js
@@ -19,7 +19,7 @@ import TaskState from '@/model/TaskState.model'
 
 /**
  * States used when the parent is stopped.
- * @type {*[]}
+ * @type {Array<TaskState>}
  */
 const isStoppedOrderedStates = [
   TaskState.SUBMIT_FAILED,
@@ -35,9 +35,9 @@ const isStoppedOrderedStates = [
 
 /**
  * Gives a single state, based on a list of states of children nodes.
- * @param childStates {Array} children nodes
- * @param isStopped {Boolean} whether the parent node is stopped or not
- * @returns {null|*} a valid Task State, or null if not found
+ * @param childStates {Array<TaskState>} children nodes
+ * @param isStopped {boolean} whether the parent node is stopped or not
+ * @returns {string} a valid Task State name, or null if not found
  * @link @see https://github.com/cylc/cylc-flow/blob/d66ae5c3ce8c749c8178d1cd53cb8c81d1560346/lib/cylc/task_state_prop.py
  */
 function extractGroupState (childStates, isStopped = false) {
