+++ tests/unit/components/cylc/tree/tree.spec.js
@@ -19,10 +19,11 @@ import CylcTree, { FAMILY_ROOT } from '@/components/cylc/tree/cylc-tree'
 import { expect } from 'chai'
 import {
   createCyclePointNode,
-  createFamilyProxyNode, createJobNode,
+  createFamilyProxyNode,
+  createJobNode,
   createTaskProxyNode,
   createWorkflowNode
-} from '@/components/cylc/tree'
+} from '@/components/cylc/tree/tree-nodes'
 import sinon from 'sinon'
 import TaskState from '@/model/TaskState.model'
